requirements.txt
asgiref==3.7.2
certifi==2024.7.4
dj-database-url==2.2.0
Django==3.2.15
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-heroku==0.3.1
django-nine==0.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==70.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
whitenoise==6.7.0
psycopg2-binary==2.9.9
celery>=5.3.0
redis>=4.5.0
